Fixed display
Upper & lower solid bounds no longer consider normal vectors when being determined. Camera setup is improved, but not working as expected. Several features were removed such as rolling the camera, planned to be added later once all of the current issues are resolved. Kind of a hack solution was added to enable orthographic projection (and it still doesn't work quite right), will tinker on this further.