Fixed rendering, added rotation
Fixed issue where the camera would seem to be slightly off-center by translating the solid to the origin and setting the camera somewhere stationary along the z-axis. Added the ability to rotate the solid around both in the x and y axis. The rotation feels slightly unnatural and difficult, this may be because it's hard to tell what's happening w/o lighting or incorrect calculations.