Fixed arcball
Arcball dragging was slow because it didn't account for the transformations in the modelview matrix when calling gluUnProject. To fix this moved gluLookAt call to alter projection matrix, then called glGetDoublev on GL_PROJECTION_MATRIX before calling gluUnProject for an updated matrix. Modelview matrix is left as identity since all the transformations are done together with the rotation.